requirements.txt
flask
flask-oauthlib
requests
sqlalchemy
